Enhancements: 1) New Task: amazon-s3-copyFile that will allow you to copy a file between buckets within S3 2) Incorporated a string (bucketName) requirement to method BucketExists to allow for validating the existence of more than one bucket. 3) Incorporated a string (fileKey) requirement to method FileExists to allow for validating the existence of multiple filenames. 4) Formatting changes on output statements
Bug Fixes:
1) Added a test to make sure a bucket exists before attempting to delete it
2) Corrected an issue where you were unable to delete files if they were inside a sub directory of the bucket
3) Added back in the CreateBucket method to the deprecated 'amazon-s3' Task to more closely mimic the original behavior of this Task

+++ S3NAntTask/S3DeleteBucketTask.cs
@@ -16,22 +16,30 @@ namespace S3NAntTask
         protected override void ExecuteTask()
         {
             Project.Log(Level.Info, "Deleting S3 bucket: {0}", BucketName);
-            using (Client)
+
+            /// Check to see if the bucket exists before we try to delete it. 
+            if (!BucketExists(BucketName))
             {
-                try
+                Project.Log(Level.Warning, "Bucket: {0} not found!", BucketName);
+                return;
+            }
+            else 
+            {
+                using (Client)
                 {
-                    DeleteBucketRequest request = new DeleteBucketRequest();
-                    request.BucketName = BucketName;
-                    //{
-                    //    BucketName = BucketName
-                    //};
-                    Client.DeleteBucket(request);
-                }
-                catch (AmazonS3Exception ex)
-                {
-                    ShowError(ex);
+                    try
+                    {
+                        DeleteBucketRequest request = new DeleteBucketRequest();
+                        request.BucketName = BucketName;
+                        Client.DeleteBucket(request);
+                    }
+                    catch (AmazonS3Exception ex)
+                    {
+                        ShowError(ex);
+                    }
                 }
             }
+            
         }
 
     }
